Tidy up reducer factories
